Process replay fixes

+++ opendbc_repo/opendbc/car/hyundai/carcontroller.py
@@ -108,8 +108,8 @@ class CarController(CarControllerBase):
                                             stopping, hud_control, actuators, CS, CC))
 
     new_actuators = actuators.as_builder()
-    new_actuators.torque = apply_torque / self.params.STEER_MAX
-    new_actuators.torqueOutputCan = apply_torque
+    new_actuators.torque = float(apply_torque / self.params.STEER_MAX)
+    new_actuators.torqueOutputCan = float(apply_torque)
     new_actuators.accel = accel
 
     self.frame += 1
